+++ packages/react/src/components/Editor/src/actions/Video/Video.tsx
@@ -2,15 +2,19 @@ import React from "react";
 import { Icon } from "@nimbus-ds/components";
 import { PlayIcon } from "@nimbus-ds/icons";
 import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
-import { UNDO_COMMAND } from "lexical";
+import { INSERT_EMBED_COMMAND } from "@lexical/react/LexicalAutoEmbedPlugin";
+
 import { Button } from "../../components";
+import { YoutubeEmbedConfig } from "../../plugins";
 
 const Video: React.FC = () => {
   const [editor] = useLexicalComposerContext();
   return (
     <Button
-      aria-label="Undo"
-      onClick={() => editor.dispatchCommand(UNDO_COMMAND, undefined)}
+      aria-label="Insert Video"
+      onClick={() =>
+        editor.dispatchCommand(INSERT_EMBED_COMMAND, YoutubeEmbedConfig.type)
+      }
     >
       <Icon source={<PlayIcon />} color="currentColor" />
     </Button>