+++ packages/react/src/components/Editor/src/nodes/YouTubeNode/YouTubeNode.tsx
@@ -1,0 +1,138 @@
+import React from "react";
+import type {
+  DOMConversionMap,
+  DOMExportOutput,
+  EditorConfig,
+  ElementFormatType,
+  LexicalEditor,
+  NodeKey,
+} from "lexical";
+import { DecoratorBlockNode } from "@lexical/react/LexicalDecoratorBlockNode";
+import { BlockWithAlignableContents } from "@lexical/react/LexicalBlockWithAlignableContents";
+
+import {
+  $createYouTubeNode,
+  convertYoutubeElement,
+} from "./youTubeNode.definitions";
+import {
+  SerializedYouTubeNode,
+  YouTubeComponentProps,
+} from "./youTubeNode.types";
+
+const YouTubeComponent: React.FC<YouTubeComponentProps> = ({
+  className,
+  format,
+  nodeKey,
+  videoID,
+}) => (
+  <BlockWithAlignableContents
+    className={className}
+    format={format}
+    nodeKey={nodeKey}
+  >
+    <iframe
+      width="560"
+      height="315"
+      src={`https://www.youtube-nocookie.com/embed/${videoID}`}
+      allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
+      allowFullScreen
+      title="YouTube video"
+    />
+  </BlockWithAlignableContents>
+);
+
+class YouTubeNode extends DecoratorBlockNode {
+  __id: string;
+
+  static getType(): string {
+    return "youtube";
+  }
+
+  static clone(node: YouTubeNode): YouTubeNode {
+    return new YouTubeNode(node.__id, node.__format, node.__key);
+  }
+
+  static importJSON(serializedNode: SerializedYouTubeNode): YouTubeNode {
+    const node = $createYouTubeNode(serializedNode.videoID);
+    node.setFormat(serializedNode.format);
+    return node;
+  }
+
+  exportJSON(): SerializedYouTubeNode {
+    return {
+      ...super.exportJSON(),
+      type: "youtube",
+      version: 1,
+      videoID: this.__id,
+    };
+  }
+
+  constructor(id: string, format?: ElementFormatType, key?: NodeKey) {
+    super(format, key);
+    this.__id = id;
+  }
+
+  exportDOM(): DOMExportOutput {
+    const element = document.createElement("iframe");
+    element.setAttribute("data-lexical-youtube", this.__id);
+    element.setAttribute("width", "560");
+    element.setAttribute("height", "315");
+    element.setAttribute(
+      "src",
+      `https://www.youtube-nocookie.com/embed/${this.__id}`
+    );
+    element.setAttribute("frameborder", "0");
+    element.setAttribute(
+      "allow",
+      "accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
+    );
+    element.setAttribute("allowfullscreen", "true");
+    element.setAttribute("title", "YouTube video");
+    return { element };
+  }
+
+  static importDOM(): DOMConversionMap | null {
+    return {
+      iframe: (domNode: HTMLElement) => {
+        if (!domNode.hasAttribute("data-lexical-youtube")) {
+          return null;
+        }
+        return {
+          conversion: convertYoutubeElement,
+          priority: 1,
+        };
+      },
+    };
+  }
+
+  // eslint-disable-next-line class-methods-use-this
+  updateDOM(): false {
+    return false;
+  }
+
+  getId(): string {
+    return this.__id;
+  }
+
+  getTextContent(): string {
+    return `https://www.youtube.com/watch?v=${this.__id}`;
+  }
+
+  decorate(_editor: LexicalEditor, config: EditorConfig): React.ReactElement {
+    const embedBlockTheme = config.theme.embedBlock || {};
+    const className = {
+      base: embedBlockTheme.base || "",
+      focus: embedBlockTheme.focus || "",
+    };
+    return (
+      <YouTubeComponent
+        className={className}
+        format={this.__format}
+        nodeKey={this.getKey()}
+        videoID={this.__id}
+      />
+    );
+  }
+}
+
+export { YouTubeNode };