+++ packages/react/src/components/Editor/src/nodes/YouTubeNode/index.ts
@@ -1,0 +1,5 @@
+import { YouTubeNode } from "./YouTubeNode";
+
+export { YouTubeNode } from "./YouTubeNode";
+export { $isYouTubeNode, $createYouTubeNode } from "./youTubeNode.definitions";
+export default YouTubeNode;