+++ packages/react/src/components/Editor/src/plugins/CustomAutoEmbedPlugin/customAutoEmbedPlugin.definitions.ts
@@ -1,0 +1,36 @@
+import { LexicalEditor } from "lexical";
+import { EmbedMatchResult } from "@lexical/react/LexicalAutoEmbedPlugin";
+
+import { PlaygroundEmbedConfig } from "./customAutoEmbedPlugin.types";
+import { INSERT_YOUTUBE_COMMAND } from "../../utils";
+
+export const YoutubeEmbedConfig: PlaygroundEmbedConfig = {
+  contentName: "Youtube Video",
+
+  exampleUrl: "https://www.youtube.com/watch?v=jNQXAC9IVRw",
+
+  insertNode: (editor: LexicalEditor, result: EmbedMatchResult) => {
+    editor.dispatchCommand(INSERT_YOUTUBE_COMMAND, result.id);
+  },
+
+  keywords: ["youtube", "video"],
+
+  // Determine if a given URL is a match and return url data.
+  parseUrl: async (url: string) => {
+    const match =
+      /^.*(youtu\.be\/|v\/|u\/\w\/|embed\/|watch\?v=|&v=)([^#&?]*).*/.exec(url);
+
+    const id = match ? (match?.[2].length === 11 ? match[2] : null) : null; // eslint-disable-line no-nested-ternary
+
+    if (id != null) {
+      return {
+        id,
+        url,
+      };
+    }
+
+    return null;
+  },
+
+  type: "youtube-video",
+};