+++ packages/react/src/components/Editor/src/plugins/CustomAutoEmbedPlugin/index.ts
@@ -1,0 +1,5 @@
+import { CustomAutoEmbedPlugin } from "./CustomAutoEmbedPlugin";
+
+export { CustomAutoEmbedPlugin } from "./CustomAutoEmbedPlugin";
+export { YoutubeEmbedConfig } from "./customAutoEmbedPlugin.definitions";
+export default CustomAutoEmbedPlugin;