+++ packages/react/src/components/Editor/src/plugins/index.ts
@@ -1,4 +1,6 @@
+export * from "./CustomAutoEmbedPlugin";
 export * from "./CustomClearFormattingPlugin";
 export * from "./CustomHeadingPlugin";
 export * from "./CustomParagraphPlugin";
+export * from "./CustomYouTubePlugin";
 export * from "./DefaultPlugins";