+++ packages/react/src/components/Editor/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./customCommands";
+export * from "./debounce";